fix(clippy): auto-fix unused imports across the workspace (21 files)
Surfaced now that the self-hosted runner is finally executing CI
(previous runs were quota-blocked and never compiled the workspace).
cargo clippy --workspace --fix --allow-dirty cleanly removed 21 unused
imports across engine-{core,browser,comms,fs,location,system,input,
social} + demo + tests.

Remaining clippy warnings (decimal-constant, manual-clamp, too-many-args,
range-empty, doc-list-indentation) are NOT in this commit — they need
small code restructuring and will land in follow-ups.

+++ engine-core/tests/security_primitives.rs
@@ -58,7 +58,7 @@ fn deadman_fires_trigger_erasure() {
 
     // Host dispatcher hands the key to erase(), receives a receipt.
     let sk = SigningKey::generate(&mut OsRng);
-    let receipt = ErasureReceipt::sign(key_id, now as i64, ErasureReason::Deadman, &sk);
+    let receipt = ErasureReceipt::sign(key_id, now, ErasureReason::Deadman, &sk);
 
     // Receipt verifies with the signing key's public half.
     receipt
